ServiceAuth token generator added
